tests/periph_spi: Convert printf specifier to PRIu32

+++ tests/periph_spi/main.c
@@ -280,7 +280,7 @@ int cmd_bench(int argc, char **argv)
     stop = xtimer_now_usec();
     sched_stop = _sched_ticks();
     printf(" 1 - write %i times %i byte:", BENCH_REDOS, 1);
-    printf("\t\t\t%i\t%"PRIu32"\n", (int)(stop - start),
+    printf("\t\t\t%"PRIu32"\t%"PRIu32"\n", (stop - start),
            xtimer_diff(sched_stop, sched_start).ticks32);
     sum += (stop - start);
     sched_sum += xtimer_usec_from_ticks(xtimer_diff(sched_stop,
@@ -296,7 +296,7 @@ int cmd_bench(int argc, char **argv)
     stop = xtimer_now_usec();
     sched_stop = _sched_ticks();
     printf(" 2 - write %i times %i byte:", BENCH_REDOS, BENCH_SMALL);
-    printf("\t\t\t%i\t%"PRIu32"\n", (int)(stop - start),
+    printf("\t\t\t%"PRIu32"\t%"PRIu32"\n", (stop - start),
            xtimer_diff(sched_stop, sched_start).ticks32);
     sum += (stop - start);
     sched_sum += xtimer_usec_from_ticks(xtimer_diff(sched_stop,
@@ -312,7 +312,7 @@ int cmd_bench(int argc, char **argv)
     stop = xtimer_now_usec();
     sched_stop = _sched_ticks();
     printf(" 3 - write %i times %i byte:", BENCH_REDOS, BENCH_LARGE);
-    printf("\t\t%i\t%"PRIu32"\n", (int)(stop - start),
+    printf("\t\t%"PRIu32"\t%"PRIu32"\n", (stop - start),
            xtimer_diff(sched_stop, sched_start).ticks32);
     sum += (stop - start);
     sched_sum += xtimer_usec_from_ticks(xtimer_diff(sched_stop,
@@ -328,7 +328,7 @@ int cmd_bench(int argc, char **argv)
     stop = xtimer_now_usec();
     sched_stop = _sched_ticks();
     printf(" 4 - write %i times %i byte to register:", BENCH_REDOS, 1);
-    printf("\t%i\t%"PRIu32"\n", (int)(stop - start),
+    printf("\t%"PRIu32"\t%"PRIu32"\n", (stop - start),
            xtimer_diff(sched_stop, sched_start).ticks32);
     sum += (stop - start);
     sched_sum += xtimer_usec_from_ticks(xtimer_diff(sched_stop,
@@ -344,7 +344,7 @@ int cmd_bench(int argc, char **argv)
     stop = xtimer_now_usec();
     sched_stop = _sched_ticks();
     printf(" 5 - write %i times %i byte to register:", BENCH_REDOS, BENCH_SMALL);
-    printf("\t%i\t%"PRIu32"\n", (int)(stop - start),
+    printf("\t%"PRIu32"\t%"PRIu32"\n", (stop - start),
            xtimer_diff(sched_stop, sched_start).ticks32);
     sum += (stop - start);
     sched_sum += xtimer_usec_from_ticks(xtimer_diff(sched_stop,
@@ -360,7 +360,7 @@ int cmd_bench(int argc, char **argv)
     stop = xtimer_now_usec();
     sched_stop = _sched_ticks();
     printf(" 6 - write %i times %i byte to register:", BENCH_REDOS, BENCH_LARGE);
-    printf("\t%i\t%"PRIu32"\n", (int)(stop - start),
+    printf("\t%"PRIu32"\t%"PRIu32"\n", (stop - start),
            xtimer_diff(sched_stop, sched_start).ticks32);
     sum += (stop - start);
     sched_sum += xtimer_usec_from_ticks(xtimer_diff(sched_stop,
@@ -376,7 +376,7 @@ int cmd_bench(int argc, char **argv)
     stop = xtimer_now_usec();
     sched_stop = _sched_ticks();
     printf(" 7 - read %i times %i byte:", BENCH_REDOS, BENCH_SMALL);
-    printf("\t\t\t%i\t%"PRIu32"\n", (int)(stop - start),
+    printf("\t\t\t%"PRIu32"\t%"PRIu32"\n", (stop - start),
            xtimer_diff(sched_stop, sched_start).ticks32);
     sum += (stop - start);
     sched_sum += xtimer_usec_from_ticks(xtimer_diff(sched_stop,
@@ -392,7 +392,7 @@ int cmd_bench(int argc, char **argv)
     stop = xtimer_now_usec();
     sched_stop = _sched_ticks();
     printf(" 8 - read %i times %i byte:", BENCH_REDOS, BENCH_LARGE);
-    printf("\t\t\t%i\t%"PRIu32"\n", (int)(stop - start),
+    printf("\t\t\t%"PRIu32"\t%"PRIu32"\n", (stop - start),
            xtimer_diff(sched_stop, sched_start).ticks32);
     sum += (stop - start);
     sched_sum += xtimer_usec_from_ticks(xtimer_diff(sched_stop,
@@ -408,7 +408,7 @@ int cmd_bench(int argc, char **argv)
     stop = xtimer_now_usec();
     sched_stop = _sched_ticks();
     printf(" 9 - read %i times %i byte from register:", BENCH_REDOS, BENCH_SMALL);
-    printf("\t%i\t%"PRIu32"\n", (int)(stop - start),
+    printf("\t%"PRIu32"\t%"PRIu32"\n", (stop - start),
            xtimer_diff(sched_stop, sched_start).ticks32);
     sum += (stop - start);
     sched_sum += xtimer_usec_from_ticks(xtimer_diff(sched_stop,
@@ -424,7 +424,7 @@ int cmd_bench(int argc, char **argv)
     stop = xtimer_now_usec();
     sched_stop = _sched_ticks();
     printf("10 - read %i times %i byte from register:", BENCH_REDOS, BENCH_LARGE);
-    printf("\t%i\t%"PRIu32"\n", (int)(stop - start),
+    printf("\t%"PRIu32"\t%"PRIu32"\n", (stop - start),
            xtimer_diff(sched_stop, sched_start).ticks32);
     sum += (stop - start);
     sched_sum += xtimer_usec_from_ticks(xtimer_diff(sched_stop,
@@ -440,7 +440,7 @@ int cmd_bench(int argc, char **argv)
     stop = xtimer_now_usec();
     sched_stop = _sched_ticks();
     printf("11 - transfer %i times %i byte:", BENCH_REDOS, BENCH_SMALL);
-    printf("\t\t%i\t%"PRIu32"\n", (int)(stop - start),
+    printf("\t\t%"PRIu32"\t%"PRIu32"\n", (stop - start),
            xtimer_diff(sched_stop, sched_start).ticks32);
     sum += (stop - start);
     sched_sum += xtimer_usec_from_ticks(xtimer_diff(sched_stop,
@@ -456,7 +456,7 @@ int cmd_bench(int argc, char **argv)
     stop = xtimer_now_usec();
     sched_stop = _sched_ticks();
     printf("12 - transfer %i times %i byte:", BENCH_REDOS, BENCH_LARGE);
-    printf("\t\t%i\t%"PRIu32"\n", (int)(stop - start),
+    printf("\t\t%"PRIu32"\t%"PRIu32"\n", (stop - start),
            xtimer_diff(sched_stop, sched_start).ticks32);
     sum += (stop - start);
     sched_sum += xtimer_usec_from_ticks(xtimer_diff(sched_stop,
@@ -472,7 +472,7 @@ int cmd_bench(int argc, char **argv)
     stop = xtimer_now_usec();
     sched_stop = _sched_ticks();
     printf("13 - transfer %i times %i byte to register:", BENCH_REDOS, BENCH_SMALL);
-    printf("\t%i\t%"PRIu32"\n", (int)(stop - start),
+    printf("\t%"PRIu32"\t%"PRIu32"\n", (stop - start),
            xtimer_diff(sched_stop, sched_start).ticks32);
     sum += (stop - start);
     sched_sum += xtimer_usec_from_ticks(xtimer_diff(sched_stop,
@@ -488,7 +488,7 @@ int cmd_bench(int argc, char **argv)
     stop = xtimer_now_usec();
     sched_stop = _sched_ticks();
     printf("14 - transfer %i times %i byte to register:", BENCH_REDOS, BENCH_LARGE);
-    printf("\t%i\t%"PRIu32"\n", (int)(stop - start),
+    printf("\t%"PRIu32"\t%"PRIu32"\n", (stop - start),
            xtimer_diff(sched_stop, sched_start).ticks32);
     sum += (stop - start);
     sched_sum += xtimer_usec_from_ticks(xtimer_diff(sched_stop,
@@ -496,7 +496,7 @@ int cmd_bench(int argc, char **argv)
 
     xtimer_sleep(1);
 
-    printf("-- - SUM:\t\t\t\t\t%i\t%"PRIu32"\n", (int)sum, sched_sum);
+    printf("-- - SUM:\t\t\t\t\t%"PRIu32"\t%"PRIu32"\n", sum, sched_sum);
 
     spi_release(spiconf.dev);
     puts("\n### All runs complete ###");
